refactore: UI layout
