+++ lib/core/routing/router_config.dart
@@ -2,17 +2,17 @@ import 'package:go_router/go_router.dart';
 import 'package:flutter/material.dart';
 import '../../presentation/screens/splash_screen.dart';
 import '../../presentation/screens/login_screen.dart';
-import '../../presentation/screens/home_screen.dart';
-import '../../presentation/screens/portfolio_screen.dart';
+import '../../presentation/screens/home/home_screen.dart';
+import '../../presentation/screens/portfolio/portfolio_screen.dart';
 import '../app/main_shell.dart';
 import '../../presentation/screens/profile_screen.dart';
-import '../../presentation/screens/transaction_history_screen.dart';
+import '../../presentation/screens/transaction_history/transaction_history_screen.dart';
 import '../../presentation/screens/add_transaction_screen.dart';
 import '../../presentation/screens/cash_screen.dart';
 import '../../presentation/screens/manage_stocks_screen.dart';
 import '../../presentation/viewmodels/cash_viewmodel.dart';
 import '../../presentation/viewmodels/manage_stocks_viewmodel.dart';
-import '../../presentation/screens/holding_details_screen.dart';
+import '../../presentation/screens/holding_details/holding_details_screen.dart';
 import '../../presentation/viewmodels/holding_details_viewmodel.dart';
 import '../../domain/entities/holding.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
