+++ lib/presentation/widgets/sticky_header_delegate.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class StickyHeaderDelegate extends SliverPersistentHeaderDelegate {
+  final double minHeight;
+  final double maxHeight;
+  final Widget child;
+
+  StickyHeaderDelegate({
+    required this.minHeight,
+    required this.maxHeight,
+    required this.child,
+  });
+
+  @override
+  Widget build(
+          BuildContext context, double shrinkOffset, bool overlapsContent) =>
+      child;
+
+  @override
+  double get maxExtent => maxHeight;
+
+  @override
+  double get minExtent => minHeight;
+
+  @override
+  bool shouldRebuild(StickyHeaderDelegate oldDelegate) =>
+      oldDelegate.minHeight != minHeight ||
+      oldDelegate.maxHeight != maxHeight ||
+      oldDelegate.child != child;
+}
