+++ lib/presentation/widgets/summary_card.dart
@@ -1,0 +1,105 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+import '../../core/constants/constants.dart';
+import '../../core/utils/formatters.dart';
+
+/// A reusable card widget for displaying financial summary information
+/// with a primary metric and two breakdown values
+class SummaryCard extends StatelessWidget {
+  final String title;
+  final double mainValue;
+  final String leftLabel;
+  final double leftValue;
+  final String rightLabel;
+  final double rightValue;
+
+  const SummaryCard({
+    super.key,
+    required this.title,
+    required this.mainValue,
+    required this.leftLabel,
+    required this.leftValue,
+    required this.rightLabel,
+    required this.rightValue,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(AppSizes.p24),
+      decoration: BoxDecoration(
+        color: AppColors.primary,
+        borderRadius: BorderRadius.circular(AppSizes.r24),
+        boxShadow: [
+          BoxShadow(
+              color: AppColors.primary.withValues(alpha: 0.3),
+              blurRadius: 20,
+              offset: const Offset(0, 10)),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(title,
+              style: GoogleFonts.inter(
+                  color: Colors.white.withValues(alpha: 0.7),
+                  fontSize: 14,
+                  fontWeight: FontWeight.w500)),
+          const SizedBox(height: 4),
+          Text(AppFormatters.formatCurrency(mainValue),
+              style: GoogleFonts.inter(
+                  color: Colors.white,
+                  fontSize: 32,
+                  fontWeight: FontWeight.bold)),
+          const SizedBox(height: AppSizes.p12),
+          // Breakdown Row
+          Row(
+            children: [
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(leftLabel,
+                        style: GoogleFonts.inter(
+                            color: Colors.white.withValues(alpha: 0.6),
+                            fontSize: 11)),
+                    const SizedBox(height: 4),
+                    Text(AppFormatters.formatCurrency(leftValue),
+                        style: GoogleFonts.inter(
+                            color: Colors.white,
+                            fontSize: 14,
+                            fontWeight: FontWeight.bold)),
+                  ],
+                ),
+              ),
+              Container(
+                width: 1,
+                height: 30,
+                color: Colors.white.withValues(alpha: 0.2),
+              ),
+              const SizedBox(width: AppSizes.p16),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(rightLabel,
+                        style: GoogleFonts.inter(
+                            color: Colors.white.withValues(alpha: 0.6),
+                            fontSize: 11)),
+                    const SizedBox(height: 4),
+                    Text(AppFormatters.formatCurrency(rightValue),
+                        style: GoogleFonts.inter(
+                            color: Colors.white,
+                            fontSize: 14,
+                            fontWeight: FontWeight.bold)),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
